Bug fix with channels and bootstrap.
The problem with the previous commit is that when bootstrap is
handled, we haven't started EventMachine yet. Since stdio
terminals don't need EM, we're okay in that case, but inserts
into other kinds of channels didn't work. Fix this by moving
the flush() for channels to after EM startup. It is ugly to
separate these two things, but works for now.