Support for ephemeral ports.
In the Bud constructor, allow either/both of the local address
and local port number to be omitted. If the address is omitted,
"localhost" is the default value ("0.0.0.0" would be another
reasonable choice, trading security for ease-of-use). If the
port is omitted, Bud will randomly search for free ports in
the range between 5000 and 25000. This is a little ugly, but
EventMachine apparently doesn't directly support ephemeral
ports.

Minor annoyance: the most common scenario is that a user will
be content with the default address, but want to pick a custom
port number. That isn't well-supported right now, because of
how default arguments in Ruby work.