Switch Bud initializer to using named parameters.
That is, Bud.new() takes a single hash specifying options to pass
to the new instance of Bud; the previously-distinct "ip" and "port"
parameters are now just possible keys for this hash.

Also, change the options hash so that it is keyed with symbols,
rather than strings. For example, @options[:disable_rewrite] rather
than @options["disable_rewrite"] -- using a symbol is more idiomatic
Ruby.

I've tried to fix as many example programs as possible, but I might
have missed a few -- should be easy to fix.