Add simple way to interact with async Bud.
Add a new method, async_do(). This takes a block and yields to that block
in the background thread at some future time. The yield is guaranteed to
happen "between" ticks, so Bud state will be internally consistent.
Because the background thread is blocked when the block is evaluated,
it can safely examine and modify Bud state (e.g., read from or insert
into collections).

Closes #36.