Fixes for new macro expansion code.
Major bug fix is that we needed to deep copy macro bodies when doing
expansion. This is because SexpProcessor mutates the Sexp tree as it processes
it; hence, if multiple subtrees in a parsetree share state, mutating one part of
the tree unintentionally mutates other areas in the tree, leading to chaos. Easy
fix is to deep copy when we expand macros; unfortunately, standard practice for
deep copying in Ruby is apparently

    Marshal.load(Marshal.dump(x))

... which is just terrible.

Other improvements: do macro expansion on the RHS of a macro body (test case
coming), add a :dump_ast Bud option, and use Sexp#sexp_type where appropriate.