Simplify implementation of bootstrap and state methods.
Now that we no longer depend on Anise, we don't need to play games with the
Module#included method. That means we can move the bootstrap and state methods
from BudModule to Module. That also means that BudModule is no longer needed!

This commit just rearranges the state and bootstrap methods; subsequent commits
will remove the usage of BudModule and BudModule itself.