Unbreak the tests.
Now that we re-eval() the body of every method defined inside a Bud class, this
surfaced a previously-hidden problem: when a Bud class was nested inside another
class or module, we were adding methods to the Bud class that contained colons
in their names. This is apparently kosher for define_method, but eval() barfes
on such method names.

Instead, only use the innermost class name for name mangling purposes.