Module improvements.
Support arbitrary methods in imported modules. Such modules will now be renamed
to be prefixed with the module's local bind name.

On a related note, make the naming scheme for bootstrap, state, and regular
instance methods more consistent. bootstrap and state methods are no longer
mangled with the import site name; all we need is the local bind name, since
instance variables are already scoped. We also now name bootstrap and state
methods using the scheme: "__state__{local bindings}__{module name}". This
matches the naming scheme used for bloom blocks, and is easier to check for
(just look for a leading "__state__" or "__bootstrap__", respectively).