Ensure that module import respects state block dependencies.
When ParseTree returns the AST for a module, any modules included by that module
are inlined into the AST. Hence, we cannot depend on inheritance information to
provide a clue to the order in which state blocks should be evaluated. (State
blocks might have dependencies because the collections in module X might be
defined with respect to a collection in a parent module of X).

To fix this, add an incrementing counter to the name of each state method; the
ID is assigned when the state method is defined (hence, the ID order must
respect the dependency order between modules). We can then sort on the value of
this ID and use that sort order when invoking state blocks.