Fixes for stratification analysis.
The major behavioral change is that we are now much more conservative about
assuming monotonicity: we only assume program fragments from a given whitelist
are monotonic, and anything we don't recognize is non-monotonic. This might
result in rejecting some legal programs, but it is the safer approach. In
particular, support for recognizing the old-style join syntax as monotonic was
removed a while ago; that should be reinstated.

Along the way, cleanup various other rewrite or strat-related bits of code.

We are now much more conservative about