Rework SIGCHLD handling in the local deployer.
This should hopefully fix the intermittent test failures. The basic problem is
that SIGCHLD handling and the wait*() family of APIs are incredibly fragile:
when you get a SIGCHLD, there is no way to know which process the SIGCHLD was
delivered for. If other parts of the program might be forking processes, we need
to spin inside the SIGCHLD handler until all possible children have been
accounted for. Unfortunately, this disturbs other parts of the program that want
to do the common "fork() + waitpid()" pattern.

Current solution is not ideal, but should be okay for now -- Closes #163.