Fix issues with localdeploy + EventMachine.
Switch the local deployment code to use EventMachine.fork_reactor, and stop the
previous practice of manually stopping EventMachine in tc_localdeploy.

In the past, this ran into some problems with EM.fork_reactor closing sockets in
the child process. That is still the current behavior, but everything seems to
work fine now.