Adjust fork- and thread-based deployment to use bootstrap blocks.
Previously the ForkDeploy and ThreadDeploy code was invoked by a deploystrap
block. deploystrap blocks are run very early during the Bud init process; this
means that the deployment infrastructure couldn't pass the IP address/port of
the deployer to the deployed children (for example).

I believe the previous coding was done to work around a suspected bug in
EventMachine (calling EM#fork_reactor causes sockets in the parent process to be
closed), but (a) I can't repro such a bug (b) the bug would cause us serious
problems anyway, so we'll need to resolve it if we run into it.