Fix signal handlers for run_fg.
The previous coding depending on being able to push() into a Queue from a signal
handler -- but if the thread that handles the signal is the same one that is
blocked trying to pop() from the queue, the pop() is apparently not awakened.

Work around this by creating a dedicated thread to handle signals. This is
overkill, but it is also safer, because it means we can usually avoid doing a
lot of work in the signal handler itself, which is generally a good idea.

Closes #174.