Add post_shutdown callbacks; use them to fix run_fg() bug.
We previously used an on_shutdown callback to wakeup the client's thread that is
blocked during a run_fg() call. This was wrong: if the client's thread is the
only active thread in the program, the program might exit before the rest of the
on_shutdown callbacks can be invoked. This was the cause of the fork-based
deployer leaving around zombies when an unhandled exception occurred (#175).

Fix this by adding the concept of a "post_shutdown" callback, which is invoked
after the Bud instance has already been shutdown. Of course, a similar problem
is possible if multiple post_shutdown callbacks have been registered, but that
seems reasonable.

Also, tweak the newly-added tc_foreground test to expose the problem.

Fixes #175.