Provide `bud_clock` as a reliable means to get the wallclock time.
Per the discussion in issue #185, using `Time.now` is unsafe because it is a
volatile value, and hence might change during the course of evaluating a
fixpoint. `bud_clock` is a new builtin method that returns a fixed wallclock
value for the duration of a single tick, similar to `current_timestamp` in
Postgres.

Fixes #185.