Fix bug in temp collections.
The previous coding inferred a schema from the RHS of a merge operation only if
the first tuple in the merge set was non-nil. Hence, if the RHS of a merge into
a temp applied a predicate such that the first tuple in the merge set was nil,
no schema would be inferred, leading to a crash shortly thereafter.