Improve deployment framework, with an eye toward auto-ft.
We previously emitted a single "node_ready" event when all the deployed nodes
had started up. This isn't a great fit for the auto-ft stuff: after a node has
failed and been restarted, we want to emit an event to acknowledge this event,
but re-emitting the single node_ready event is wrong/confusing.

Instead, emit a single node_ready event whenever a node becomes ready, and fixup
the places that make use of node_ready. We can then change auto-ft to simply
emit a node_ready event whenever a node is recovered, which seems sensible.

Along the way, fixup the fork and thread-based deployment test cases to work for
non-default values of NUM_DEPLOY_FORKS and NUM_DEPLOY_THREADS, respectively.