Improve Bud error reporting.
If an error occurs inside an EventMachine callback, check whether it is a
BudError (or subclass); if so, don't print the full ugly stacktrace, since most
BudErrors include sufficient error reporting information already.

This essentially fixes #205. We still produce a stacktrace when an error occurs
during Bud initialization, but that will be less easy to improve: we want to
raise an exception, which Ruby handles by dumping a stack trace.