Enable a conservative form of message batching.
If multiple tuples are delivered in the same EM::Connection#receive_data call,
those tuples will be processed together in a batch. At least in the limited
testing I've done, this doesn't actually happen very often (and batch sizes are
typically small), but it is at least possible.

More aggressive batching schemes are possible, although we'll need to decide
about how to manage the tradeoff between latency and throughput.