Clarify argmin and argmax behavior in case of ties.
argmin and argmax might produce multiple tuples. That is:

    result <= t.argmin([], :a)

If multiple tuples contain the minimal value of `a`, argmin will produce both
result tuples. Document this behavior, and add a test for it. (The only real
alternative would be to produce an error in this case -- that might also be
worth considering, but it seems less flexible than the current behavior.)