Monkeypatch Enumerable#pro as an alias for map.
The map => pro rewrite is overly eager: it rewrites some cases in which map is
applied to a Ruby Enumerable that is not a BudCollection. It would be possible
to fix this by making the rewrite more accurate, but it seems hard to guarantee
accuracy without being overly conservative. Instead, define Enumerable#pro as an
alias for map.

Fixes #225.