Avoid hanging if an error occurs during sync_callback.
The calling thread will be blocked trying to pop from a queue, and the previous
behavior would leave that thread hanging. Use the `on_shutdown` facility to
wakeup all such threads; `sync_callback` will now raise an exception. To enable
this, improve the `on_shutdown` facility to allow shutdown callbacks to be
canceled.

The need to use two different callbacks (for the failure and success cases) is
unfortunate, but I'm not sure there's a good way around it in the short term. A
possible improvement would be add a parameter to `register_callback`: if true,
the callback should be invoked on runtime shutdown.

Fixes #228.