Enable EventMachine for Bud single-step mode.
Previously, we didn't start EventMachine when Bud was started using the tick()
method ("single-stepped mode"). This meant that Bud functionality that depends
on EM was broken for this mode, including channels and timers.

Thus, reimplement tick() as follows:

 * On the first call to tick(), we start EM and bind to a local UDP socket,
   unless run_bg/run_fg has already been called.

 * Events received via EventMachine (inbound network messages and timer events)
   will be queued; when tick() is called, all outstanding events will be
   consumed in a single fixpoint.

 * To close the UDP server socket, stop_bg() must be called. "stop_bg()" is no
   longer a good name; we should probably rename it to "shutdown()", or simply
   "stop()". Since stop_bg() can (and should) be used for single-stepped Bud
   instances, the close_tables() method is redundant and has been removed.

 * It is supported to create an instance of Bud, call tick() one or more times
   on it, and then call run_bg() or run_fg(). A new method pause_bg() can be
   used to switch from "run asynchronously in the background" mode to
   single-stepped mode.

Along the way, I noticed a problem with the timer implementation: the unique IDs
associated with timer tuples were generated when the periodic table was
defined. Hence every firing of a given periodic would share the same "unique"
ID. This caused problems particularly for timers used with tick(), since
multiple firings of the same timer can now be consumed in the same fixpoint.

Note that the previous "lazy" mode (needed for REBL) and the new functionality
added in this commit slightly overlap; I'll check with Bill to see if we can
eliminate lazy mode.