Add a notion of "scratch lattices".
By default, the value in a lattice persists over Bud timesteps; this makes the
normal lattice behavior analogous to "table" collections.

However, a "scratch-like" lattice behavior could also be useful, for two
reasons:

  (a) Lattices can be used to achieve monotonicity within a timestep; persisting
      values and/or achieving monotonicity/confluence across timesteps might not
      be necessary.

  (b) Scratch lattices can still be used in confluent programs, provided that
      all inputs to the lattice are the result of a chain of monotonic queries
      that bottom out in persistent collections (either tables or persistent
      lattices). For a program of this form, which collections and which
      lattices to persist can be made a view materialization (optimization)
      question, rather than something the user needs to decide.

Lattices are persistent by default; scratch lattices can be declared as follows:

  lat_foo :name, :scratch => true

As with a scratch collection, a scratch lattice is reset at the start of each
timestep; user-defined lattices must provide a #reset method to achieve this.