Tweak @pending implementation for lattices.
Rather than storing pending (@next) tuples in a list (i.e., set lattice),
instead store them in another instance of the lattice itself. Merging via <+
does a merge with the "pending lattice"; at the end of every timestep, we merge
the pending lattice with the main lattice, and reset the pending lattice.