Require ruby2ruby < 1.3.1.
Ruby2Ruby 1.3.1 has a bug that breaks the `tc_wc` test case: ruby2ruby removes
empty hash arguments to method calls, which is not a sane/safe thing to do in
general. I reported a ruby2ruby bug
(https://github.com/seattlerb/ruby2ruby/issues/3), but in the meantime, this
commit changes Bud to depend on and load a version of Ruby2Ruby < 1.3.1.