Remove a few of the unspeakable hacks from the viz code.
Previously, the viz code used a bunch of ad hoc and mostly wrong rules to
determine if a table is a builtin table or not: e.g., tables with names starting
with "t_" or ending with "_tbl" were ignored in various places. That resulted in
ignoring legitimate user-defined tables with those names; it also missed builtin
tables that didn't match the pattern.

Replace that scheme with something more reasonable: Bud now builds a hash
containing the builtin collections and we just consult that hash to see if a
collection is a builtin or not.