Fix a subtle bug in MergeMapLattice.
Because <= is (currently) destructive, we need to take care to make a defensive
copy of the value associated with a key when it is seen for the first
time. Otherwise subsequent updates to the value in the MergeLattice might also
affect the source of the value, which would be wrong. Perhaps the cleaner
long-term solution is to get rid of the destructive (over-writing) behavior of
the <= operator.

Also, implement MergeMapLattice#reveal.