Quick hacks needed for causal delivery protocol.
Key point is that if m is a MergeMapLattice and m.keys does not contain, we want
m[k] to return the "bottom" value for the domain of map values. That is,
m[k].lt_eq(x) should be true for any value x in the domain of map values.