Fixes for running passing tc_temp's tests. Two problems: (i) schema may not be available during initialization time (ii) items pushed towards a join lacked tuple_accessors of their respective sources
Plus, a temporary fix for examples/paths.rb
A temp table name of ":j" doesn't work correctly because of namespace pollution.
The name "j" refers to an integer in a bloom stmt "j <= ... ", not a table.
The temporary fix is to rename it :jj. Separate correct fix for this issue.

	modified:   examples/basics/paths.rb
	modified:   lib/bud/collections.rb
	modified:   lib/bud/executor/elements.rb