Local variable bindings captured before "eval" cause interference. For example,
  @rewritten_strata.each_with_index do |src_ary,i|
      ...
      eval ...

In this block, the eval will see names like src_ary and i. Which means,
a bloom statement cannot refer to a table called "src_ary" or "i".

The solution is to capture the variable bindings on entry to Bud::initialize,
so that other local bindings do not interfere. This may not be fool-proof, because
var bindings from outer blocks may still interfere, although I have not been
able to conclusively come up with an example that breaks it.

	modified:   examples/basics/paths.rb
	modified:   lib/bud.rb