Reject duplicate import names found across included modules.
We previously rejected attempts to use the same import bind name by a single
module/class; we now also reject situations in which module/class X includes Y
and Z, and both Y and Z import a module with the same bind name. There are a
situations in which this pattern wouldn't cause problems, but it seems safer to
be conservative and disallow it.