Further fixes for include and import rewrite interactions.
The basic problem is that a module's import table only contains entries for the
direct imports of the module. The table should actually contain entries for the
module and for any modules included by that module. This could be fixed in two
ways:

  1. Hook into the `include` system and merge the included module's import table
     into the import table at the include site.

  2. Before we use the import table, iterate over any included modules and merge
     their import tables into the parent import table.

I avoided #1, because hacking the `include` system has been troublesome in the
past. Hence, this commit implements #2 -- more specifically, #2 was implemented
in a prior commit (5cdc95f772b1ab866e654233cf0801498b00d6fa), and this commit
makes sure we perform #2 before the other circumstances in which the import
table is used.