First step towards porting to 1.9.x and JRuby, and simplifying rewriting.
ParseTree for Ruby 1.8.7 could figure out the ast of a proc or
method. The library does not work in JRuby or Ruby 1.9.3. We now rely
on the pure-ruby ruby_parser library that can translate source ruby
text to AST. monkeypatch uses Kernel#caller to figure out the source
file/line of bloom blocks. However, this technique does not work for
the current module import strategy which relies on examining all
methods of all modules related to the bud_instance. Since a module can
be split over multiple source files, and since there is no correlation
between a module nameand a source file name (unlike Java), we cannot
quickly figure out the source for a particular call.

However, this technique does not work for the current module import
strategy which relies on examining all methods of all modules related
to the bud_instance. Since a module can be split over multiple source
files, and since there is no correlation between a module nameand a
source file name (unlike Java), we cannot quickly figure out the
source for a particular call. Besides, there are several bits of eval'd code
that do not have lines in a source file.

For this reason, module import has been disabled in this version. The next step
is to eliminate the need to rewrite module references at all. This will be
done by having an object encapsulate an imported module.

	modified:   lib/bud.rb
	modified:   lib/bud/bud_meta.rb
	modified:   lib/bud/collections.rb
	modified:   lib/bud/executor/elements.rb
	modified:   lib/bud/monkeypatch.rb
	modified:   lib/bud/rewrite.rb
	new file:   lib/bud/source.rb