Allow duplicate import bind names.
We previously rejected programs that use the same bind name for more than one
import statement. Instead, allow such programs; we implement them by merging all
import tables associated with the same bind name. This can lead to problems,
because the code in modules with the same bind name can interact: e.g., if two
modules declare a table with the same name and are imported with the same bind
name, we'll try to create two tables with the same name and fail. But since this
behavior can also be useful, seems better to allow it.

Fix #258.