Perf tweak for collections.
The previous implementation of #merge invoked #uniq on the input buffer (if it
was supported). This is unnecessary: if the buffer does contain duplicates, they
will be ignored by the merge process anyway. It is also slightly wrong: if the
collection schema consists only of key columns, the consistent behavior would be
to raise a key violation. Finally, it hurts performance: removing the #uniq call
improves the performance of the scratch_derive benchmark by about 5%.