Remove wrapper pointer from Bud::Lattice values.
Keeping a point from values back to the wrapper that holds the value was always
a hack, so dispense with it. Now, a reference to a lattice collection name will
yield the lattice's wrapper; this ensures that "<=" and similar operators
continue to work. To get the current lattice value, the #current_value method
must be invoked explicitly. This is clumsy -- a subsequent checkin will enable
rewriting so that references to lattice collection names (e.g., "x") on the RHS
of rules are replaced with "x.current_value".