Fix a subtle bug in lattice merges for collections.
The previous coding searched [@storage, @delta, @new_delta] looking for
conflicting tuples, in that order; hence, if multiple delta tuples conflict with
the same @storage tuple, we'd merge each delta tuple against the storage tuple
-- but then throw away the previously-merged tuple.

The fix is to search for matches by order of "recency": @new_delta, @delta, and
then @storage. In general, this is a fairly fragile coding pattern.