Improve exception reporting.
When an exception occurs Bud evaluation, we trap and re-raise it (typically with
extra information about which Bloom rule caused the problem). The downside is
that this replaces the exception's backtrace with the site of the trap/reraise
logic.

As a quick hack, improve the situation by including the first line of the
original backtrace in the exception message; in many cases, that should provide
enough information to identify the problem.

See #251 for the general issue.