Remove BudCollection#method_missing.
This was used to proxy calls to undefined methods on Bud collections down to the
underlying collection storage (e.g., a hash). In practice, only two methods were
actually used in the test suite (#length and #exists?), so it seems to define
individual methods on an as-needed basis. This is safer (simple proxying might
not do the right thing), and it also makes some analysis for the lattice stuff a
bit harder.