Avoid using "y" as a collection name in tc_module.
This causes a conflict with the "y" installed into Kernel by the "yaml" gem
(which is loaded by the BUST code). This isn't a problem with the current coding
(because the "y" collection is nested inside a module that is name-mangled by
the import system), but it is a foot gun and should be removed.

This is just a port of some of the changes from
aea6357fe32e15c479e0b34ac5b9041f636fa200 by Sriram; thanks to him for noticing
and debugging the problem.