Invalidation support is more correct (passes tc_collection, tc_aggs at the least).
"Invalidation" means tuples are erased from storage, and cached state is removed where appropriate.

Invalidation means slightly different things in the different cases:

1. BudTable: Tables are invalidated on negation. Likewise with all persistent tables.

2. BudScratch: "Source" scratches (those that are only in the rhs, never in the lhs) are always considered invalidated at the beginning of a tick. All other scratches are lazily invalidated (storage nullified) only when tuples are pending to be added, which means that to maintain scratch semantics, the storage has to be cleared before the new tuples are added.

3. BudChannel/BudTerminal: These have different behavior for the incoming and outgoing. The storage for incoming tuples (@storage) is always invalidated at the beginning
of a tick. Outgoing storage (@output_storage) is a cache that sticks around until an upstream table has been invalidated.

Once a table is invalidated, it requests its dependents (lhs of a rule is "dependent" on the collections on the rhs) to remove cached/memoized state. The reason is that the table wil shortly be scanned fully, and all downstream dependent tables and elements will be repopulated. Cache invalidation is a transitive operation. However the transitivity stops at a BudTable .. see comment in BudTable:invalidate_cache.

----------------------------------------------------------------------

Also in this commit ..

a) added BudCollection::tick_deltas, an array of changes to storage during thetick. For higher strata, this is the list to scan for all the changes to the table. If the table is invalidated, we revert to scanning the entire storage as before.

b) We only examine application defined tables (bud.rb: @app_tables) in every tick, not the built-in-tables as they do not change.

	modified:   ../lib/bud.rb
	modified:   ../lib/bud/bud_meta.rb
	modified:   ../lib/bud/collections.rb
	modified:   ../lib/bud/executor/elements.rb
	modified:   tc_collections.rb