Add a "sealed" lattice type.
The basic idea is that you can seal another lattice value by applying a
(non-morphic) function. The sealed value is not allowed to change; subsequent
code can then safely invoke non-morphic methods on the sealed value without that
code being flagged as non-monotonic.

Not 100% baked yet: the fact that sealing is always viewed as non-monotonic
seems weird, but this is at least a start toward something that might have
value.