The concept of "delta" is overloaded. In SN evaluation, delta is only what is _produced_ incrementally, but there is no notion of increments to the EDB. In Bud, user code can add tuples to delta between ticks.
With this commit, pending always goes to delta. Since any table can possibly get deltas from outside (in a sync_do block, for example) and not just bud.merge_targets, we add the list of collections scanned in each stratum to the set of merge_targets. The definition of "merge_targets" is changed from the earlier definition "the set of tables in this stratum on the lhs", to "the set of tables in this stratum that
may have deltas to be merged".

	modified:   ../lib/bud.rb
	modified:   ../lib/bud/collections.rb