DBM is now ready for incremental processing
tc_dbm modified to standardize on <+ instead of a mixture of <= and <<

In particular, << should not be used outside bootstrap, because it modifies
storage directly.

Collections get an extra bootstrap method. This is to ensure that both pending gets moved into deltas. Inside a bootstrap block and in code outside tick, <+ and <= behave identically.

	modified:   ../lib/bud.rb
	modified:   ../lib/bud/collections.rb
	modified:   ../lib/bud/monkeypatch.rb
	modified:   ../lib/bud/storage/dbm.rb
	modified:   tc_collections.rb
	modified:   tc_dbm.rb