Replace SumLattice with PositiveSetLattice.
This is more general-purpose: if we restrict the set to only allowing
non-negative inputs, we can easily support "sum" as a morphism. It can also be
implemented in only a few lines of code by sub-classing SetLattice.

Note that the previous SetLattice used nonces to basically support bag semantics
for the input to the sum (i.e., SQL SUM(), not SUM(DISTINCT)), whereas the new
version does duplicate elimination.