Changes to correct for NTT benchmark problems, but this commit is still wrong; it is overly conservative for joins, and ends up unnecessarily invalidating both branches of a join.
The core problem is this:
kvresponse <= (kvget * kvstate)

which translates to the following wiring:
  kvget/Scanner   --->
                        * join --->   kvresponse
  kvstate/Scanner --->

kvget is an interface, so it is invalidated on tick, which in turn invalidates both the join and kvresponse. In the current invalidation scheme, kvstate's scanner notices that the join is invalidated, and marks itself for a rescan, which messes the join's partial invalidation mechanism

The solution is for each node to propagate its rescan requests up to its source nodes. That's the work for the next commit.

	modified:   lib/bud.rb
	modified:   lib/bud/collections.rb
	modified:   lib/bud/executor/elements.rb
	modified:   lib/bud/executor/join.rb
	modified:   test/tc_joins.rb