Correction to previous commit. Lot more focus on rescan, as opposed to invalidation. Used to be that a node set its invalidation requirements based on the invalidation state of others. In this commit, a node sets its invalidation/rescan requirements based on others' rescan state. Also, a node can now nominate other nodes for rescan.
We've come a full circle. Started with a node being able to nominate others for invalidation, then each node being only responsible for its own nominations based on others' state, then back to a node responsible for nominating others for rescan (not for invalidation). This is logical (in hindsight) because as invalidations go top-down, rescans requests bubble up.

This commit runs the NTT benchmark correctly and in O(1) time.

	modified:   lib/bud/collections.rb
	modified:   lib/bud/executor/elements.rb
	modified:   lib/bud/executor/join.rb
	modified:   test/tc_aggs.rb