Workaround Ruby2Ruby bug.
On the RHS of a rule, we previously operated on an AST rooted with s(:arglist),
since Ruby considers a Bloom rule to be a function call (invoking <=/<+/etc. on
the lhs identifier and passing it the rhs as arguments).

An AST rooted with s(:arglist) is not really sensible, and causes Ruby2Ruby
1.3.0 to misbehave -- e.g., given s(:arglist, s(:hash, ...)), it ignores the
hash AST element.