Remove support for :signal_handling == :bloom.
This was buggy (e.g., when delivering signals as Bud tuples, it would only
deliver to one of the Bud instances in the process -- which might not even be
running any more). The implementation was also dubious: doing Bud#sync_do from
inside a signal handler is almost surely unsafe. It was also causing
non-deterministic test failures in other parts of the test suite, because of the
tick_all_instances() call performed by the signal check timer.

Philosophically, I would prefer to keep Bud signal handling as simple as
possible. If there is a compelling need for this feature it can be reintroduced,
but it should be done carefully.