Renable stripping the outer Sexp from rule RHS.
This is always s(:arglist). An AST rooted in s(:arglist) is not really sensible;
it also causes some versions of Ruby2Ruby to misbehave in some situations (in
particular, when we want the RHS of a rule to be a hash literal, which is used
by the lattice branch).