Issue #271. notin error fixed.
PushNotIn element used to depend upon PushSHJoin, and then compute an anti-join at end of stratum. Not any more. It is an error to wait until stratum_end to output tuples, since the fixpoint loop is already done. Also, notin is not really a symmetric join, since stratification ensures that the rhs is already sealed in a lower stratum, and will not change in the current stratum. For these two reasons, PushNotIn has been simplified.

"notin" has been special-cased in rewrite as noted in the issue.

And as a simplification unrelated to issue 271, input buffering in join has been removed.

	modified:   lib/bud/executor/elements.rb
	modified:   lib/bud/executor/join.rb
	modified:   lib/bud/rewrite.rb