Fix equality semantics for lattice objects.
It seems reasonable that two lattices that represent the same value should
compare as equal (e.g., Bud::MaxLattice.new(5) == Bud::MaxLattice.new(5)). To
achieve this, implement #==(o) based on calling reveal().

Note that actually testing for equality between two lattice values in a Bloom
program is non-monotonic in general, but providing more sensible equality
semantics is useful elsewhere (in particular, in the implementation of the lpair
lattice).