Significantly revise and simplify "implicit fold" behavior.
In the push-based runtime, we already split sets of tuples into individual
tuples for push-based processing. Thus, we need only convert individual tuples
into singleton lattice elements (which there was already code to do) and merge
each resulting lattice element as it is seen. This avoids the need to wrap and
unwrap lattice values from inside collections, which was slow, ugly, and
confusing.