Fix a major bug in the rescan logic.
The previous coding neglected to properly propagate rescan requirements to the
source nodes of a node that has been invalidated. For example, consider this
program:

  s2 <= t1
  s2 <= s1

where s1, s2 are scratches and t1 is a table. Since s2 is invalidated on every
tick, we need to mark t1 for rescan to repopulate (part of) s2. There was code
that intended to propagate the rescan flag in the right manner, but it was
buggy.

It is worth noting that the cause of the bug was a misuse of the "+=" operator
provided by the Set (and Array) classes. This operator is *not* side-effecting.
For example, with "x" as a Set or Array instance:

  x += [1,2,3]

is equivalent to:

  x = x + [1,2,3]

Obviously, this does *not* update the original object referenced by "x" -- "x"
is just reassigned to reference the newly formed set/array. So for example,

  def foo(x); x += [1,2,3]; end

foo() is a no-op. Using Set#merge or Array#concat yields the intended behavior.

This coding pattern occurs in a few other places, which I will fix in a separate
commit. I didn't see any obvious bugs but using += does an extra copy which is
usually unnecessary (since the intent is usually to update the input in place).