Fix a major bug in the join rescan logic.
The previous coding for rescanning join operators had two different
bugs. Consider a program like:

  s2 <= s1
  s2 <= (t1 * t2).pairs { ... }

where s1, s2 are scratches and t1, t2 are tables. Since s2 is invalidated, the
join of (t1*t2) needs to be rescanned. After a recent fix (see commit
be9b3106d5229397b0fee6fc6ca3570b561482c9), the rescan flag should be correctly
propagated to the join operator. However, the join operator only consulted the
rescan flag during PushSHJoin#insert() -- i.e., when new tuples arrived on
either of the join inputs. As far as I can tell that is entirely the wrong
place: arrival of new join inputs is not related to the need to reproduce the
join's output. Hence, this commit moves that logic to PushSHJoin#flush(), which
you can think of as a workaround for the fact that we don't have scanners over
join operators.

This fix exposed another bug: replay_join() used the wrong indexes when calling
process_matches() to produce join outputs, which meant that either the wrong
output would be produced or an exception would be raised. Prior to this commit
replay_join() was not exercised by the test suite (!), which is why the coding
mistake wasn't caught.