Performance optimization for simple projection rules.
If a rule feeds the content of one collection into another without applying a
projection/selection operation, we can implement this via just a ScannerElement,
not a ScannerElement => PushElement. That is, rules like:

  x <= y

Avoiding the extra dataflow element yields a small performance improvement (~10%
for a microbenchmark that moves data between seven scratches in this manner),
but since this pattern is reasonably common, it seems worth the trouble.