Fix incorrect notin output.
Again, the notin operator didn't implement the rescan logic correctly; as a
result, it failed to correctly rescan when connected to a scratch collection,
for example.

PushNotIn *attempted* to define a #stratum_end() method, but the nesting of the
Ruby `end` statements was wrong, which meant we ended up defining
Bud#stratum_end by accident :) However, the previous stratum_end() code was
wrong anyway, so it didn't really matter.