Add lmap#apply_morph and lmap#apply_monotone.
These methods apply a morphism or a monotone function, respectively, to all the
values in the lmap. The return value is an lmap with the same key set but a
transformed value set. Note that the key associated with a given value is not
currently passed into the function that is invoked by the two apply methods.