Re-enable support for each_with_index() over Bud collections.
This code was mistakenly deleted (!) in commit
2beb7e82b97c6547adc0784cfada338630fe50b4, but it was never covered by a test
case (applying each_with_index to the result of a Bud expression such as `sort`
always worked, and was covered by the test suite). Note that applying
`each_with_index` directly to a collection is probably a bad idea anyway, since
the iteration order (and hence assigned indices) is not well-defined.