Fix a minor bug introduced in the lattice branch.
When feeding tuples that originated in one Bud collection into another, we
incorrectly raised an error if (a) the target collection had one or more key
columns (b) the input tuple did not contain a column at the corresponding
offset. e.g., feeding [5] into a collection with keys [:a, :b] would have raised
an error; the correct (or at least consistent) behavior is to nil-pad the input
tuple and not raise an error.